refactor: Moved class to another file and used lambda expr to simplify code
